Very late first commit on a Brainfuck interpreter for Chip8.
Uses a python script to build an assembly file which
we then use rchip8 to assemble and run like normal.

The program is inserted in the image as ascii chars,
beginning at 0x202 (due to a jump instruction). It
ends with a temrinating value of 0. Giving you a max
of 252 characters in your program.

The data section begins at 0x300 (program is padded
to ensure this). There are 256 bytes of data you can
access.

For now the input command is non functional
(I plan to add an input section like data and prog).
Also the output command will stop the interpreter
after putting a single byte into v0.

This was enough to be able to subtract 2 numbers:
+++++>+++[-<->]<.

And see the result in V0 in the debug output.